working tree no redux
